refactor(pipeline): piece-level predicates get their own module (#439)
Pure relocation. No behavior change, no ledger entry, no release-log
bullet -- the differential is byte-identical because no parse moves.

Ten objects leave _group.py verbatim, comment blocks included: the
rules.md citations live in those comments and have to travel with the
code they cite. _is_title_piece, _PERIOD_ABBREV, _is_leading_title,
_leading_titles, _is_suffix_piece, _segment_holds_no_name, Peel,
_peel_walk, _trailing_start, _peel_trailing.

_vocab.py was the other candidate and rules itself out in its own
docstring -- "Text-level tests used by more than one stage;
token/piece-level predicates live with their stage." That contract is
deliberate (is_wholly_suffix documents two departures from it), so the
piece layer gets its own module rather than bending the text layer's.
The split is by what the question TAKES, which is the rule
mechanisms.md#ONE-PREDICATE-PER-QUESTION states.

_trailing_start travels with the peel although only _peel_walk and
_peel_trailing cross a stage boundary: decisions.md describes the S2
peel as those three together, and splitting a documented unit across
modules to satisfy a usage rule would be worse than carrying one
group-only function along.

Five predicates stay in _group, being group's alone: _is_prefix_piece,
_is_conj_piece, _is_rootname, _is_maiden_marker_piece, and the local
closures. group imports back what it still uses; nothing imports
_pieces from below, so there is no cycle.

Three couplings the move had to carry, none of which grep would have
found from the function names alone:

- rules.md H2, H3 and S2 name the module their citation sits in, and
  test_doc_citations asserts set EQUALITY against citing modules, so
  their implemented: lines move to _pieces.py.
- test_regex_sync tracks _PERIOD_ABBREV by module path in three places
  (the import, the assertion, and the declared-copy roster).
- test_layering iterates ALLOWED.items(), so a module absent from that
  dict is never checked AT ALL. _pieces.py is registered with a
  tighter contract than the generic stage allowance -- _state and
  _vocab only -- which makes the module docstring's claim executable.
  Verified non-vacuous: the entry rejects an added nameparser._render
  import on a scratch copy.
